Remove numpy dtype dependence of smooth function.
The problem was that dtype=int arrays where kept as int despite
the fact that the smoothed values would be floats. With a arr.astype(float)
line within the function, the smoothing now works with float numbers.

+++ scan/data_processing.py
@@ -160,7 +160,7 @@ def smooth(
 
     # ensure array is (t,d,s) shaped for the algorithm
     x_data_shape = x_data.shape
-    smoothed = x_data.copy()
+    smoothed = x_data.copy().astype(float)
     if smoothed.ndim == 1:
         smoothed = smoothed[:, None, None]
     elif smoothed.ndim == 2:
